修复了进度条显示异常（int改long）

+++ app/src/main/java/db/ThreadDAO.java
@@ -11,7 +11,7 @@ import stl.ThreadInfo;
 public interface ThreadDAO {
     public void insertThread(ThreadInfo threadInfo);
     public void deleteThread(String url);
-    public void updateThread(String url,int thread_id,int finished);
+    public void updateThread(String url,int thread_id,long finished);
     public List<ThreadInfo>getThreads(String url);
     public boolean isExists(String url,int thread_id);
 }
